pyproject.toml
[tool.poetry]
name = "mlops-basics"
version = "0.1.0"
description = ""
authors = ["froukje <falbrechtg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
datasets = "^2.18.0"
transformers = "^4.39.1"
torch = {version = "2.0.1+cu118", source = "torch"}
pytorch-lightning = "^2.2.1"
scikit-learn = "^1.4.1.post1"
tensorboard = "^2.16.2"
wandb = "^0.16.5"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
hydra-core = "^1.3.2"
hydra-colorlog = "^1.2.0"


[tool.poetry.group.formatting.dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"



[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
